fix: can't create user
userEntity's userAccount does not need @JsonIgnore, only UserAccountEntity's user need
